+++ packages/shared/src/ISplitOperation.ts
@@ -1,0 +1,7 @@
+export interface ISplitOperation {
+  /** The current status of the operation. */
+  status: string;
+
+  /** How complete the operation is, from 0-100. */
+  percentage: number;
+}
