+++ packages/shared/src/IStreamState.ts
@@ -1,0 +1,9 @@
+import { ChannelTuning } from "./ChannelTuning";
+
+export interface IStreamState {
+  /** The explicit ID of this singleton. Should always be 0. */
+  _id: number;
+
+  /** The tunings for each channel on the stream. */
+  tunings: ChannelTuning[];
+}
