+++ packages/shared/src/index.ts
@@ -1,5 +1,9 @@
 export * from "./ChannelIndex";
 export * from "./ChannelTuning";
+export * from "./ISplitOperation";
+export * from "./IStreamState";
+export * from "./ITimestamps";
+export * from "./ITransmission";
 export * from "./StreamMessage";
 export * from "./TransmissionStem";
 export * from "./WebSocketMessage";
